backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9  # Keep for Alembic migrations (sync operations)
aioredis==2.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-decouple==3.8
python-json-logger==2.0.7
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
alembic==1.13.1
weasyprint==60.2
celery[redis]==5.3.4
redis>=4.5.2,<5.0.0
httpx==0.25.2
python-multipart==0.0.6 